Utf8IteratorRange: explicitly cast values before shifting to avoid involvement of unobvious integer promotion

+++ stingraykit/string/Utf8IteratorRange.h
@@ -48,19 +48,19 @@ namespace stingray
 
 			const u8 c1 = (u8)*it++;
 			if (c0 >= 0xc2 && c0 <= 0xdf)
-				return ((c0 & 0x1f) << 6) | (c1 & 0x3f);
+				return ((u32)(c0 & 0x1f) << 6) | (c1 & 0x3f);
 
 			STINGRAYKIT_CHECK(it != _end, MalformedUtf8Exception());
 
 			const u8 c2 = (u8)*it++;
 			if (c0 >= 0xe0 && c0 <= 0xef)
-				return ((c0 & 0x0f) << 12) | ((c1 & 0x3f) << 6) | (c2 & 0x3f);
+				return ((u32)(c0 & 0x0f) << 12) | ((u32)(c1 & 0x3f) << 6) | (c2 & 0x3f);
 
 			STINGRAYKIT_CHECK(it != _end, MalformedUtf8Exception());
 
 			const u8 c3 = (u8)*it++;
 			if (c0 >= 0xf0 && c0 <= 0xf4)
-				return ((c0 & 0x07) << 18) | ((c1 & 0x3f) << 12) | ((c2 & 0x3f) << 6) | (c3 & 0x3f);
+				return ((u32)(c0 & 0x07) << 18) | ((u32)(c1 & 0x3f) << 12) | ((u32)(c2 & 0x3f) << 6) | (c3 & 0x3f);
 
 			STINGRAYKIT_THROW(MalformedUtf8Exception());
 		}
